requirements.txt
anthropic>=0.34.0
openai>=1.0.0
google-generativeai>=0.8.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
apscheduler>=3.10.0
jinja2>=3.1.0
python-dotenv>=1.0.0
